requirements.txt
flask
flask-restful
requests
beautifulsoup4
pillow
pytesseract
python-dotenv